requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1

# Database
psycopg2-binary==2.9.7

# Authentication
Authlib==1.2.1
requests==2.31.0

# Content processing
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
newspaper3k==0.2.8

# AI/LLM integration
openai==0.28.1

# Payment processing
stripe==6.6.0

# Email
Flask-Mail==0.9.1

# Text-to-speech
gTTS==2.3.2

# Task scheduling
APScheduler==3.10.4

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
validators==0.22.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0

# Production server
gunicorn==21.2.0